chore: Add ErrorPage component and update index.js
This commit adds the ErrorPage component to the components directory and updates the index.js file to export the ErrorPage component. The ErrorPage component will be used to display an error page when there is an issue with loading a page. This change improves the user experience by providing a clear error message and a way to navigate back to the previous page.

Please review and merge the changes.